requirements.txt
semantic-kernel
python-dotenv
openai
